Clean-up in Employment.conf
The config parameter Employment-GUID-generation-attributes was removed from
the example config file since it's not used.

Employment-type-attribute and Employment-type-value were removed both from
the example config file and from the source code since it's redundant (we
already choose which users are Teachers when we configure the StudentGroup
and its relation to Teacher). It was also less flexible since the attribute
determining who is a teacher had to have a specific value.

Misc clean-up in the comments in the example config file.