The base scenario and base configuration updated with more details.
All users now have civic number in LDAP and the students have a school year.

One of the groups is now a "Klass", and this is now determined with an
attribute instead of being hard coded in the config.

Civic number is not used in the base configuration since it would normally
not be sent with the EGIL profile.