Adds ExpectErrors attribute in test specifications.
This makes it possible to tell the test framework that we expect the
client to print something to stderr.

The test framework will only forward it to the console if we didn't
expect any errors. Also, if we do expect errors but the client is silent
the test is considered a failing test.