Stricter handling of bad UUIDs.
By default we will now give a warning if an object is read in with
a bad UUID (upper case letters or bad syntax).

The object will however by default be allowed to be loaded so as
to not break existing deployments.

The warning can be disabled, and there is also a new config variable
for saying that objects with bad UUIDs should not be loaded.

Fixes #163